+++ src/test/java/io/lettuce/core/FastShutdown.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2016 the original author or authors.
+ * Copyright 2011-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,7 +30,7 @@ public class FastShutdown {
      * @param redisClient
      */
     public static void shutdown(AbstractRedisClient redisClient) {
-        redisClient.shutdown(10, 10, TimeUnit.MILLISECONDS);
+        redisClient.shutdown(0, 10, TimeUnit.MILLISECONDS);
     }
 
     /**
@@ -39,6 +39,6 @@ public class FastShutdown {
      * @param clientResources
      */
     public static void shutdown(ClientResources clientResources) {
-        clientResources.shutdown(10, 10, TimeUnit.MILLISECONDS);
+        clientResources.shutdown(0, 10, TimeUnit.MILLISECONDS);
     }
 }